+++ packages/paste-core/components/account-switcher/src/useAccountSwitcher.ts
@@ -1,0 +1,1 @@
+export {useMenuState as useAccountSwitcherState} from '@twilio-paste/menu';