+++ packages/paste-core/components/product-switcher/src/useProductSwitcher.ts
@@ -1,0 +1,1 @@
+export {useMenuState as useProductSwitcherState} from '@twilio-paste/menu';