+++ packages/paste-core/components/sidebar/src/SidebarHeader.tsx
@@ -15,7 +15,8 @@ export const SidebarHeader = React.forwardRef<HTMLButtonElement, SidebarHeaderPr
         {...safelySpreadBoxProps(props)}
         element={element}
         ref={ref}
-        padding="space60"
+        paddingX="space60"
+        minHeight="sizeTopbar"
         borderBottomWidth="borderWidth10"
         borderBottomStyle="solid"
         borderBottomColor="colorBorderInverseWeaker"