+++ packages/paste-core/components/topbar/src/user-dialog/UserDialogUserInfo.tsx
@@ -1,0 +1,49 @@
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import {Box, safelySpreadBoxProps} from '@twilio-paste/box';
+import {MediaObject, MediaFigure, MediaBody} from '@twilio-paste/media-object';
+import {Separator} from '@twilio-paste/separator';
+import {Avatar} from '@twilio-paste/avatar';
+
+import type {UserDialogUserInfoProps} from '../types';
+import {UserDialogContext} from './UserDialogContainer';
+
+export const UserDialogUserInfo = React.forwardRef<HTMLDivElement, UserDialogUserInfoProps>(
+  ({children, element = 'USER_DIALOG', ...props}, ref) => {
+    const {avatarProps} = React.useContext(UserDialogContext);
+    return (
+      <>
+        <Box
+          {...safelySpreadBoxProps(props)}
+          backgroundColor="colorBackgroundWeak"
+          padding="space50"
+          borderTopLeftRadius="borderRadius30"
+          borderTopRightRadius="borderRadius30"
+          element={`${element}_USER_INFO`}
+          ref={ref}
+        >
+          <MediaObject as="div" verticalAlign="center">
+            <MediaFigure as="div" spacing="space50">
+              <Avatar
+                variant="user"
+                size="sizeIcon80"
+                name={avatarProps.name}
+                src={avatarProps.src}
+                icon={avatarProps.icon}
+                element={`${element}_USER_AVATAR`}
+              />
+            </MediaFigure>
+            <MediaBody as="div">{children}</MediaBody>
+          </MediaObject>
+        </Box>
+        <Separator orientation="horizontal" />
+      </>
+    );
+  }
+);
+UserDialogUserInfo.displayName = 'UserDialogUserInfo';
+
+UserDialogUserInfo.propTypes = {
+  children: PropTypes.node.isRequired,
+  element: PropTypes.string,
+};